% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pirna_removal.R
\name{get_pirna_distances}
\alias{get_pirna_distances}
\title{Get Exact Hamming Distances for all piRNAs}
\usage{
get_pirna_distances(dna_seq)
}
\arguments{
\item{dna_seq}{Character string; the input DNA sequence.}
}
\value{
A numeric vector of the minimum Hamming distances for each known piRNA.
}
\description{
Slices the input sequence into 20-mer sliding windows and computes the
minimum Hamming distance between any 20-mer in the sequence and the
established C. elegans piRNA database.
}
