% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glo_optimization.R
\name{optimize_glo}
\alias{optimize_glo}
\title{Germline Optimize (GLO) a DNA Sequence}
\usage{
optimize_glo(dna_seq)
}
\arguments{
\item{dna_seq}{Character string; the input DNA sequence to optimize.}
}
\value{
A character string containing the GLO-optimized DNA sequence.
}
\description{
Implementation of GLO algorithm developed by Daniel Dickinson (https://pubmed.ncbi.nlm.nih.gov/30109984/).
}
\details{
Employs a greedy, localized perturbation algorithm to optimize a coding sequence
for germline expression based on a massive 12-mer fitness dictionary. To achieve
high performance, this algorithm utilizes a base-R implementation of Binary Search
against a pre-sorted dictionary, reducing lookup complexity from O(N) to O(log N).

Furthermore, when evaluating a codon mutation, it does not recalculate the fitness
of the entire gene. It maps and recalculates only the 4 overlapping 12-mer windows
directly impacted by the substitution, radically improving iteration speed.
}
