% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimize_transgene.R
\name{probabilistic_recode}
\alias{probabilistic_recode}
\title{Probabilistic DNA Recoder}
\usage{
probabilistic_recode(dna_seq, usage_column_name)
}
\arguments{
\item{dna_seq}{Character string; the DNA sequence to recode.}

\item{usage_column_name}{Character string; the name of the column in the internal codon frequency table to use.}
}
\value{
A character string of the recoded DNA sequence.
}
\description{
Recodes a DNA sequence by substituting synonymous codons according to a
specified usage frequency column in the internal codon table.
}
