% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(launch.browser = TRUE),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for `shinyAppObj`, since in the `shinyAppDir` case, a
`global.R` file can be used for this purpose.}

\item{options}{Named options that should be passed to the `runApp` call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode").}

\item{enableBookmarking}{Can be one of "url", "server", or "disable".}

\item{uiPattern}{A regular expression that will be applied to each `GET` request
to determine whether the `ui` should be used to handle the request.}

\item{...}{Additional arguments to pass to `golem_opts`.
See `?golem::get_golem_options` for more details.}
}
\description{
Initializes and launches the WormTransgeneBuilder Shiny application.
This function acts as the primary entry point for the package.
}
