% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_codon_optimization.R
\name{sample_new_codon}
\alias{sample_new_codon}
\title{Sample a Synonymous Codon}
\usage{
sample_new_codon(aa, old_codon, usage_column_name)
}
\arguments{
\item{aa}{Character string; the single-letter amino acid code.}

\item{old_codon}{Character string; the 3-nucleotide codon to be replaced.}

\item{usage_column_name}{Character string; the column name in the internal codon frequency table to use for weighted sampling.}
}
\value{
A character string of the newly sampled 3-nucleotide codon.
}
\description{
Selects a different synonymous codon for a given amino acid based on the
specified usage frequencies. Used primarily to mutate sequences when
destroying restriction enzyme sites or piRNA targets.
}
